requirements.txt
opencv-python
face_recognition==1.3.0
opencv-python-headless==4.8.0.76
numpy==1.24.4
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
